requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.2
python-multipart==0.0.6
pydantic==1.10.12
PyYAML==6.0.1
